Adding the screen renderer
To make sure that the rendering code is extracted from the window management code, the screen class will be used to render all entities and screens. It takes the graphics object used at render time and will render all components to it. It is implemented in the game window and is currently using a demonstrative random pixel art generation as a proof of concept until entities are added.