Creating entity
This is the template for entities which are anything that should be drawn to the screen. These will eventually be implemented by the characters and pieces of the level.

I've also added two utility functions to the screen which are for rendering sprites in a nicer fashion, hiding the Graphics2D implementation.