Adding levels
Levels are, as the name would suggest, the levels that the game is played on. They contain a background sprite and a set of entities which are designed to represent components that are part of the level such as platforms and items. It just contains utility functions for quickly adding and removing entities