Updating rendering code and adding testing details
Removing the test of the emitter system which reset the title and making the run function update the title with the FPS. The render code has been updated to limit it to 200 FPS at most which reduces the murderous speed and prevents the crashing of machines running it. The fullscreen setting has been removed for now just to make testing easier. The actual render code has been updated in line with the JavaDoc to make sure that it renders properly and does not miss frames and buffers