Adding mouse handler
The mouse handler acts as a buffer joining together all three of the mouse listeners. If does very little processing on its own, just handing off the event to any registered listeners. There is a listener interface that contains all of the listener functions and an abstract implementation which means that only the wanted functions need to be overwritten when using an anonymous class.