Adding media player controller
This is a runnable designed to play audio files. This will be used to implement the sound system. Because it is standalone and a runnable, I should be able to run multiple sounds at once. This is not yet tested but it is currently assumed. Once the audio manager is added, it will be tested