Making entities obey basic gravity and implementing update calls
Ive added the missing update calls so they now flow from the GameState to the Level and Entities and from the Level to their entities. This means that all entities should receive updates and only the GameState needs to be registered.
I've also exposed access to the level bounds which allows me to implement collisions for the entities which I've used to make them fall until they hit something. This works reasonably well but it is a little bit jumpy and annoying but this will be fixed later on