Implementing jumping mechanics
This is very basic right now. The jumping will basically just make you keep going up for a fixed amount of time and then let gravity work out the rest. You can only jump if you are not currently jumping and are on a fixed surface. The speeds are a bit fucked right now, mainly because I'm just doing it as offsets rather than velocities which I may implement after this but that is more a thought for later