Adding projectile implementation
This is a testing implementation right now (hence loading the image on every launched projectile) but is a proof of concept right now. It has a rate limited projectile launch system which summons one in the same direction from the top left of the creature. This shows that the projectiles are not accurate representations of projectiles as they do not follow the correct parabolic arc but they slightly arc which is enough to satisfy me. This velocity system needs to be implemented for jumping but that will be coming soon